+++ app/src/main/java/com/umc/approval/data/repository/upload/UploadRepository.kt
@@ -1,8 +1,9 @@
 package com.umc.approval.data.repository.upload
 
-import com.umc.approval.data.dto.community.TalkUploadDto
+import com.umc.approval.data.dto.upload.post.TalkUploadDto
 import com.umc.approval.data.dto.profile.CategoriesDto
-import com.umc.approval.data.dto.upload.UploadDto
+import com.umc.approval.data.dto.upload.post.ApprovalUploadDto
+import com.umc.approval.data.dto.upload.post.ReportUploadDto
 import com.umc.approval.data.retrofit.instance.RetrofitInstance.uploadApi
 import okhttp3.ResponseBody
 import retrofit2.Call
@@ -18,12 +19,17 @@ class UploadRepository() {
     }
 
     /**approval upload api*/
-    fun upload_approval_item(accessToken: String, upload: UploadDto):Call<ResponseBody> {
+    fun upload_approval_item(accessToken: String, upload: ApprovalUploadDto):Call<ResponseBody> {
         return uploadApi.upload_approval(accessToken, upload)
     }
 
-    /**approval upload api*/
-    fun upload_talk_item(accessToken: String, talkUpload: TalkUploadDto):Call<ResponseBody> {
-        return uploadApi.upload_community_talk(accessToken, talkUpload)
+    /**tok upload api*/
+    fun upload_tok_item(accessToken: String, toktok: TalkUploadDto):Call<ResponseBody> {
+        return uploadApi.upload_community_tok(accessToken, toktok)
+    }
+
+    /**report upload api*/
+    fun upload_report_item(accessToken: String, report: ReportUploadDto):Call<ResponseBody> {
+        return uploadApi.upload_community_report(accessToken, report)
     }
 }
