+++ app/src/main/java/com/umc/approval/data/retrofit/api/UploadAPI.kt
@@ -1,14 +1,15 @@
 package com.umc.approval.data.retrofit.api
 
-import com.umc.approval.data.dto.community.TalkUploadDto
+import com.umc.approval.data.dto.upload.post.TalkUploadDto
 import com.umc.approval.data.dto.profile.CategoriesDto
-import com.umc.approval.data.dto.upload.UploadDto
+import com.umc.approval.data.dto.upload.post.ApprovalUploadDto
+import com.umc.approval.data.dto.upload.post.ReportUploadDto
 import okhttp3.ResponseBody
 import retrofit2.Call
 import retrofit2.http.*
 
 /**
- * Upload and Community Upload API
+ * Approval Upload and Community Upload API
  * */
 interface UploadAPI {
 
@@ -22,19 +23,31 @@ interface UploadAPI {
     @Headers("content-type: application/json")
     fun upload_approval(
         @Header("Authorization") accessToken: String,
-        @Query("upload") upload: UploadDto):Call<ResponseBody>
+        @Query("upload") upload: ApprovalUploadDto):Call<ResponseBody>
 
     /**
      * @Post
      * accessToken : 사용자 검증 토큰
      * upload : body, category, images, tags, voteItems, opengraph 정보
-     * talk upload api
      * */
-    @POST("/upload/community/talk")
+    @POST("/community/toktoks")
     @Headers("content-type: application/json")
-    fun upload_community_talk(
+    fun upload_community_tok(
         @Header("Authorization") accessToken: String,
-        @Query("talkUpload") talkUpload: TalkUploadDto):Call<ResponseBody>
+        @Query("toktok") toktok: TalkUploadDto
+    ):Call<ResponseBody>
+
+    /**
+     * @Post
+     * accessToken : 사용자 검증 토큰
+     * upload : body, category, images, tags, voteItems, opengraph 정보
+     * */
+    @POST("/community/reports")
+    @Headers("content-type: application/json")
+    fun upload_community_report(
+        @Header("Authorization") accessToken: String,
+        @Query("report") report: ReportUploadDto
+    ):Call<ResponseBody>
 
     /**
      * @Post
